+++ packages/expo-modules-core/ios/Tests/ConvertiblesSpec.swift
@@ -363,6 +363,22 @@ class ConvertiblesSpec: ExpoSpec {
         let transparent = try CGColor.convert(from: "transparent", appContext: appContext)
         expect(transparent.alpha) == .zero
       }
+      
+      it("converts from PlatformColor") {
+        let color = try CGColor.convert(from: ["semantic": ["invalid_color", "systemRed", "systemBlue"]], appContext: appContext)
+        expect(color) == UIColor.systemRed.cgColor
+      }
+      
+      it("converts from DynamicColorIOS") {
+        let color = try CGColor.convert(from: ["dynamic": ["light": "#000", "dark": ["semantic": "systemGray"]]], appContext: appContext)
+        testColorComponents(color, 0x00, 0x00, 0x00, 0xFF)
+      }
+      
+      it("converts from DynamicColorIOS with traits") {
+        let color = try UIColor.convert(from: ["dynamic": ["light": "#000", "dark": ["semantic": "systemGray"]]], appContext: appContext)
+        let traits = UITraitCollection(userInterfaceStyle: .dark)
+        expect(color.resolvedColor(with: traits)) == UIColor.systemGray.resolvedColor(with: traits)
+      }
 
       it("throws when string is invalid") {
         testInvalidHexColor("")